+++ husky_components/toggler/main.js
@@ -24,7 +24,6 @@ define([], function() {
 
     var defaults = {
             instanceName: 'undefined',
-            checked: false,
             hiddenCheckbox: true,
             outline: false
         },
@@ -41,13 +40,13 @@ define([], function() {
          * namespace for events
          * @type {string}
          */
-            eventNamespace = 'husky.toggler.',
+        eventNamespace = 'husky.toggler.',
 
         /**
          * raised after initialization process
          * @event husky.toggler.<instance-name>.initialize
          */
-            INITIALIZED = function() {
+        INITIALIZED = function() {
             return createEventName.call(this, 'initialized');
         },
 
@@ -56,7 +55,7 @@ define([], function() {
          * @event husky.toggler.<instance-name>.changed
          * @param {boolean} on True if button is on
          */
-            CHANGED = function() {
+        CHANGED = function() {
             return createEventName.call(this, 'changed');
         },
 
@@ -65,12 +64,12 @@ define([], function() {
          * @event husky.toggler.<instance-name>.change
          * @param {boolean} on To turn the button on or off
          */
-            CHANGE = function() {
+        CHANGE = function() {
             return createEventName.call(this, 'change');
         },
 
         /** returns normalized event names */
-            createEventName = function(postFix) {
+        createEventName = function(postFix) {
             return eventNamespace + (this.options.instanceName ? this.options.instanceName + '.' : '') + postFix;
         };
 
@@ -84,7 +83,6 @@ define([], function() {
             this.options = this.sandbox.util.extend(true, {}, defaults, this.options);
 
             this.hasId = false;
-            this.checked = !!this.options.checked;
             this.$checkbox = null;
             this.$wrapper = null;
 
@@ -92,7 +90,9 @@ define([], function() {
             this.bindDomEvents();
             this.bindCustomEvents();
 
-            this.setData();
+            if (this.$el.data('checked') === undefined) {
+                this.setData(false);
+            }
 
             this.sandbox.emit(INITIALIZED.call(this));
         },
@@ -113,16 +113,16 @@ define([], function() {
             }
 
             //wrapper to insert all content into
-            this.$wrapper = this.sandbox.dom.createElement('<div class="'+ constants.wrapperClass +'"/>');
+            this.$wrapper = this.sandbox.dom.createElement('<div class="' + constants.wrapperClass + '"/>');
             this.sandbox.dom.html(this.$el, this.$wrapper);
 
             this.generateHiddenCheckbox();
 
-            if (this.checked === true) {
+            if (this.getData() === true) {
                 this.setChecked(false);
             }
 
-            this.sandbox.dom.append(this.$wrapper, this.sandbox.dom.createElement('<div class="'+ constants.switchClass +'"/>'));
+            this.sandbox.dom.append(this.$wrapper, this.sandbox.dom.createElement('<div class="' + constants.switchClass + '"/>'));
         },
 
         /**
@@ -130,12 +130,12 @@ define([], function() {
          */
         bindDomEvents: function() {
             this.sandbox.dom.on(this.$el, 'click', function() {
-               this.toggleButton();
+                this.toggleButton();
             }.bind(this));
 
             //if toggler has id enable clicks on labels
             if (this.hasId === true) {
-                this.sandbox.dom.on('label[for="'+ this.options.instanceName +'"]', 'click', function() {
+                this.sandbox.dom.on('label[for="' + this.options.instanceName + '"]', 'click', function() {
                     this.toggleButton();
                 }.bind(this));
             }
@@ -146,7 +146,7 @@ define([], function() {
          */
         bindCustomEvents: function() {
             this.sandbox.on(CHANGE.call(this), function(checked) {
-                if (this.checked !== checked) {
+                if (this.getData() !== checked) {
                     if (checked === true) {
                         this.setChecked(true);
                     } else {
@@ -160,7 +160,7 @@ define([], function() {
          * Toggles the button state
          */
         toggleButton: function() {
-            if (this.checked === true) {
+            if (this.getData() === true) {
                 this.unsetChecked(true);
             } else {
                 this.setChecked(true);
@@ -172,9 +172,8 @@ define([], function() {
          */
         setChecked: function(emit) {
             this.sandbox.dom.addClass(this.$el, constants.checkedClass);
-            this.checked = true;
-            this.setData();
-            this.updateCheckbox(this.checked);
+            this.setData(true);
+            this.updateCheckbox(this.getData());
 
             if (emit !== false) {
                 this.sandbox.emit(CHANGED.call(this), true);
@@ -186,9 +185,8 @@ define([], function() {
          */
         unsetChecked: function(emit) {
             this.sandbox.dom.removeClass(this.$el, constants.checkedClass);
-            this.checked = false;
-            this.setData();
-            this.updateCheckbox(this.checked);
+            this.setData(false);
+            this.updateCheckbox(this.getData());
 
             if (emit !== false) {
                 this.sandbox.emit(CHANGED.call(this), false);
@@ -200,7 +198,7 @@ define([], function() {
          */
         generateHiddenCheckbox: function() {
             if (this.options.hiddenCheckbox === true) {
-                this.$checkbox = this.sandbox.dom.createElement('<input type="checkbox" name="'+ this.options.instanceName +'"/>');
+                this.$checkbox = this.sandbox.dom.createElement('<input type="checkbox" name="' + this.options.instanceName + '"/>');
                 this.sandbox.dom.append(this.$wrapper, this.$checkbox);
             }
         },
@@ -218,9 +216,16 @@ define([], function() {
         /**
          * Sets the data-checked attribute for the toggler
          */
-        setData: function() {
-            this.sandbox.dom.data(this.$el, 'checked', (!!this.checked) ? 'checked' : '');
+        setData: function(isChecked) {
+            this.$el.data('checked', !!isChecked);
+        },
+
+        /**
+         * Returns the data-checked attribute for the toggler
+         * @returns {boolean}
+         */
+        getData: function() {
+            return this.$el.data('checked');
         }
     };
-
 });