+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # CHANGELOG for husky
 
 * dev-develop
+    * ENHANCEMENT #593 Changed toggler to only use data attribute
     * FEATURE     #578 Adjusted design for multiple select
     * FEATURE     #586 Enhanced table view